requirements.txt
fastapi
uvicorn[standard]
jinja2
python-dotenv
google-generativeai
aiofiles
pillow
asyncpg
